+++ packages/cloud_functions/cloud_functions/android/src/main/java/io/flutter/plugins/firebase/functions/FlutterFirebaseFunctionsPlugin.java
@@ -18,7 +18,6 @@ import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
-import io.flutter.plugin.common.PluginRegistry.Registrar;
 import io.flutter.plugins.firebase.core.FlutterFirebasePlugin;
 import java.io.IOException;
 import java.io.InterruptedIOException;
@@ -43,12 +42,6 @@ public class FlutterFirebaseFunctionsPlugin
    * <p>Use this when adding the plugin to your FlutterEngine
    */
   public FlutterFirebaseFunctionsPlugin() {}
-
-  public static void registerWith(Registrar registrar) {
-    final MethodChannel channel = new MethodChannel(registrar.messenger(), METHOD_CHANNEL_NAME);
-    final FlutterFirebaseFunctionsPlugin plugin = new FlutterFirebaseFunctionsPlugin(channel);
-    channel.setMethodCallHandler(plugin);
-  }
 
   @Override
   public void onAttachedToEngine(@NonNull FlutterPluginBinding binding) {