+++ packages/firebase_analytics/firebase_analytics/android/src/main/java/io/flutter/plugins/firebase/analytics/Constants.java
@@ -1,0 +1,13 @@
+package io.flutter.plugins.firebase.analytics;
+
+public class Constants {
+  public static final String AD_STORAGE_CONSENT_GRANTED = "adStorageConsentGranted";
+  public static final String ANALYTICS_STORAGE_CONSENT_GRANTED = "analyticsStorageConsentGranted";
+  public static final String USER_ID = "userId";
+  public static final String EVENT_NAME = "eventName";
+  public static final String PARAMETERS = "parameters";
+  public static final String ENABLED = "enabled";
+  public static final String MILLISECONDS = "milliseconds";
+  public static final String NAME = "name";
+  public static final String VALUE = "value";
+}