+++ packages/firebase_analytics/firebase_analytics/example/lib/main.dart
@@ -1,4 +1,3 @@
-// ignore_for_file: require_trailing_commas
 // Copyright 2017 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
@@ -8,22 +7,26 @@ import 'dart:async';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_analytics/firebase_analytics.dart';
-import 'package:firebase_analytics/observer.dart';
 
 import 'tabs_page.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  await Firebase.initializeApp();
+  await Firebase.initializeApp(
+    options: const FirebaseOptions(
+      apiKey: 'AIzaSyAHAsf51D0A407EklG1bs-5wA7EbyfNFg0',
+      appId: '1:448618578101:ios:2bc5c1fe2ec336f8ac3efc',
+      messagingSenderId: '448618578101',
+      projectId: 'react-native-firebase-testing',
+    ),
+  );
   runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
   const MyApp({Key? key}) : super(key: key);
 
-  static FirebaseAnalytics analytics = FirebaseAnalytics();
-  static FirebaseAnalyticsObserver observer =
-      FirebaseAnalyticsObserver(analytics: analytics);
+  static FirebaseAnalytics analytics = FirebaseAnalytics.instance;
 
   @override
   Widget build(BuildContext context) {
@@ -32,11 +35,9 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      navigatorObservers: <NavigatorObserver>[observer],
       home: MyHomePage(
         title: 'Firebase Analytics Demo',
         analytics: analytics,
-        observer: observer,
       ),
     );
   }
@@ -47,12 +48,10 @@ class MyHomePage extends StatefulWidget {
     Key? key,
     required this.title,
     required this.analytics,
-    required this.observer,
   }) : super(key: key);
 
   final String title;
   final FirebaseAnalytics analytics;
-  final FirebaseAnalyticsObserver observer;
 
   @override
   _MyHomePageState createState() => _MyHomePageState();
@@ -75,8 +74,8 @@ class _MyHomePageState extends State<MyHomePage> {
         'int': 42,
         'long': 12345678910,
         'double': 42.0,
-// Only strings and numbers (ints & doubles) are supported for GA custom event parameters:
-// https://developers.google.com/analytics/devguides/collection/analyticsjs/custom-dims-mets#overview
+        // Only strings and numbers (ints & doubles) are supported for GA custom event parameters:
+        // https://developers.google.com/analytics/devguides/collection/analyticsjs/custom-dims-mets#overview
         'bool': true.toString(),
       },
     );
@@ -84,7 +83,7 @@ class _MyHomePageState extends State<MyHomePage> {
   }
 
   Future<void> _testSetUserId() async {
-    await widget.analytics.setUserId('some-user');
+    await widget.analytics.setUserId(id: 'some-user');
     setMessage('setUserId succeeded');
   }
 
@@ -103,7 +102,8 @@ class _MyHomePageState extends State<MyHomePage> {
   }
 
   Future<void> _testSetSessionTimeoutDuration() async {
-    await widget.analytics.android?.setSessionTimeoutDuration(2000000);
+    await widget.analytics
+        .setSessionTimeoutDuration(const Duration(milliseconds: 20000));
     setMessage('setSessionTimeoutDuration succeeded');
   }
 
@@ -112,76 +112,62 @@ class _MyHomePageState extends State<MyHomePage> {
     setMessage('setUserProperty succeeded');
   }
 
+  AnalyticsEventItem itemCreator() {
+    return AnalyticsEventItem(
+      affiliation: 'affil',
+      coupon: 'coup',
+      creativeName: 'creativeName',
+      creativeSlot: 'creativeSlot',
+      discount: 2.22,
+      index: 3,
+      itemBrand: 'itemBrand',
+      itemCategory: 'itemCategory',
+      itemCategory2: 'itemCategory2',
+      itemCategory3: 'itemCategory3',
+      itemCategory4: 'itemCategory4',
+      itemCategory5: 'itemCategory5',
+      itemId: 'itemId',
+      itemListId: 'itemListId',
+      itemListName: 'itemListName',
+      itemName: 'itemName',
+      itemVariant: 'itemVariant',
+      locationId: 'locationId',
+      price: 9.99,
+      currency: 'USD',
+      promotionId: 'promotionId',
+      promotionName: 'promotionName',
+      quantity: 1,
+    );
+  }
+
   Future<void> _testAllEventTypes() async {
     await widget.analytics.logAddPaymentInfo();
     await widget.analytics.logAddToCart(
       currency: 'USD',
       value: 123,
-      itemId: 'test item id',
-      itemName: 'test item name',
-      itemCategory: 'test item category',
-      quantity: 5,
-      price: 24,
-      origin: 'test origin',
-      itemLocationId: 'test location id',
-      destination: 'test destination',
-      startDate: '2015-09-14',
-      endDate: '2015-09-17',
+      items: [itemCreator(), itemCreator()],
     );
-    await widget.analytics.logAddToWishlist(
-      itemId: 'test item id',
-      itemName: 'test item name',
-      itemCategory: 'test item category',
-      quantity: 5,
-      price: 24,
-      value: 123,
-      currency: 'USD',
-      itemLocationId: 'test location id',
-    );
+    await widget.analytics.logAddToWishlist();
     await widget.analytics.logAppOpen();
     await widget.analytics.logBeginCheckout(
       value: 123,
       currency: 'USD',
-      transactionId: 'test tx id',
-      numberOfNights: 2,
-      numberOfRooms: 3,
-      numberOfPassengers: 4,
-      origin: 'test origin',
-      destination: 'test destination',
-      startDate: '2015-09-14',
-      endDate: '2015-09-17',
-      travelClass: 'test travel class',
+      items: [itemCreator(), itemCreator()],
     );
     await widget.analytics.logCampaignDetails(
-      source: 'test source',
-      medium: 'test medium',
-      campaign: 'test campaign',
-      term: 'test term',
-      content: 'test content',
-      aclid: 'test aclid',
-      cp1: 'test cp1',
+      source: 'source',
+      medium: 'medium',
+      campaign: 'campaign',
+      term: 'term',
+      content: 'content',
+      aclid: 'aclid',
+      cp1: 'cp1',
     );
     await widget.analytics.logEarnVirtualCurrency(
       virtualCurrencyName: 'bitcoin',
       value: 345.66,
     );
-    await widget.analytics.logEcommercePurchase(
-      currency: 'USD',
-      value: 432.45,
-      transactionId: 'test tx id',
-      tax: 3.45,
-      shipping: 5.67,
-      coupon: 'test coupon',
-      location: 'test location',
-      numberOfNights: 3,
-      numberOfRooms: 4,
-      numberOfPassengers: 5,
-      origin: 'test origin',
-      destination: 'test destination',
-      startDate: '2015-09-13',
-      endDate: '2015-09-14',
-      travelClass: 'test travel class',
-    );
+
     await widget.analytics.logGenerateLead(
       currency: 'USD',
       value: 123.45,
@@ -193,27 +179,14 @@ class _MyHomePageState extends State<MyHomePage> {
       level: 5,
       character: 'witch doctor',
     );
-    await widget.analytics.logLogin();
+    await widget.analytics.logLogin(loginMethod: 'sign up');
     await widget.analytics.logPostScore(
       score: 1000000,
       level: 70,
       character: 'tiefling cleric',
     );
-    await widget.analytics.logPresentOffer(
-      itemId: 'test item id',
-      itemName: 'test item name',
-      itemCategory: 'test item category',
-      quantity: 6,
-      price: 3.45,
-      value: 67.8,
-      currency: 'USD',
-      itemLocationId: 'test item location id',
-    );
-    await widget.analytics.logPurchaseRefund(
-      currency: 'USD',
-      value: 45.67,
-      transactionId: 'test tx id',
-    );
+    await widget.analytics
+        .logPurchase(currency: 'USD', transactionId: 'transaction-id');
     await widget.analytics.logSearch(
       searchTerm: 'hotel',
       numberOfNights: 2,
@@ -229,10 +202,30 @@ class _MyHomePageState extends State<MyHomePage> {
       contentType: 'test content type',
       itemId: 'test item id',
     );
+    await widget.analytics.logSelectPromotion(
+      creativeName: 'promotion name',
+      creativeSlot: 'promotion slot',
+      items: [itemCreator()],
+      locationId: 'United States',
+    );
+    await widget.analytics.logSelectItem(
+      items: [itemCreator(), itemCreator()],
+      itemListName: 't-shirt',
+      itemListId: '1234',
+    );
+    await widget.analytics.logScreenView(
+      screenName: 'tabs-page',
+    );
+    await widget.analytics.logViewCart(
+      currency: 'USD',
+      value: 123,
+      items: [itemCreator(), itemCreator()],
+    );
     await widget.analytics.logShare(
-        contentType: 'test content type',
-        itemId: 'test item id',
-        method: 'facebook');
+      contentType: 'test content type',
+      itemId: 'test item id',
+      method: 'facebook',
+    );
     await widget.analytics.logSignUp(
       signUpMethod: 'test sign up method',
     );
@@ -241,31 +234,31 @@ class _MyHomePageState extends State<MyHomePage> {
       virtualCurrencyName: 'bitcoin',
       value: 34,
     );
+    await widget.analytics.logViewPromotion(
+      creativeName: 'promotion name',
+      creativeSlot: 'promotion slot',
+      items: [itemCreator()],
+      locationId: 'United States',
+      promotionId: '1234',
+      promotionName: 'big sale',
+    );
+    await widget.analytics.logRefund(
+      currency: 'USD',
+      value: 123,
+      items: [itemCreator(), itemCreator()],
+    );
     await widget.analytics.logTutorialBegin();
     await widget.analytics.logTutorialComplete();
     await widget.analytics.logUnlockAchievement(id: 'all Firebase API covered');
     await widget.analytics.logViewItem(
-      itemId: 'test item id',
-      itemName: 'test item name',
-      itemCategory: 'test item category',
-      itemLocationId: 'test item location id',
-      price: 3.45,
-      quantity: 6,
-      currency: 'USD',
-      value: 67.8,
-      flightNumber: 'test flight number',
-      numberOfPassengers: 3,
-      numberOfRooms: 1,
-      numberOfNights: 2,
-      origin: 'test origin',
-      destination: 'test destination',
-      startDate: '2015-09-14',
-      endDate: '2015-09-15',
-      searchTerm: 'test search term',
-      travelClass: 'test travel class',
+      currency: 'usd',
+      value: 1000,
+      items: [itemCreator()],
     );
     await widget.analytics.logViewItemList(
-      itemCategory: 'test item category',
+      itemListId: 't-shirt-4321',
+      itemListName: 'green t-shirt',
+      items: [itemCreator()],
     );
     await widget.analytics.logViewSearchResults(
       searchTerm: 'test search term',
@@ -309,17 +302,22 @@ class _MyHomePageState extends State<MyHomePage> {
             onPressed: _testSetUserProperty,
             child: const Text('Test setUserProperty'),
           ),
-          Text(_message,
-              style: const TextStyle(color: Color.fromARGB(255, 0, 155, 0))),
+          Text(
+            _message,
+            style: const TextStyle(color: Color.fromARGB(255, 0, 155, 0)),
+          ),
         ],
       ),
       floatingActionButton: FloatingActionButton(
         onPressed: () {
-          Navigator.of(context).push(MaterialPageRoute<TabsPage>(
+          Navigator.of(context).push(
+            MaterialPageRoute<TabsPage>(
               settings: const RouteSettings(name: TabsPage.routeName),
               builder: (BuildContext context) {
-                return TabsPage(widget.observer);
-              }));
+                return const TabsPage();
+              },
+            ),
+          );
         },
         child: const Icon(Icons.tab),
       ),