+++ packages/firebase_analytics/firebase_analytics_web/lib/firebase_analytics_web.dart
@@ -1,17 +1,30 @@
-// ignore_for_file: require_trailing_commas
-import 'package:firebase/firebase.dart' as firebase;
+// Copyright 2021 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 import 'package:firebase_analytics_platform_interface/firebase_analytics_platform_interface.dart';
+import 'package:firebase_analytics_web/utils/exception.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:firebase_core_web/firebase_core_web_interop.dart'
+    as core_interop;
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
-import 'package:meta/meta.dart';
 
-/// Web implementation for [FirebaseAnalyticsPlatform]
+import 'interop/analytics.dart' as analytics_interop;
+
+/// Web implementation of [FirebaseAnalyticsPlatform]
 class FirebaseAnalyticsWeb extends FirebaseAnalyticsPlatform {
-  /// A constructor that allows tests to override the firebase.Analytics object.
-  FirebaseAnalyticsWeb({@visibleForTesting firebase.Analytics? analytics})
-      : _analytics = analytics ?? firebase.analytics();
+  /// instance of Analytics from the web plugin
+  analytics_interop.Analytics? _webAnalytics;
 
-  /// Instance of Analytics from the web plugin.
-  final firebase.Analytics _analytics;
+  /// Lazily initialize [_webAnalytics] on first method call
+  analytics_interop.Analytics get _delegate {
+    return _webAnalytics ??=
+        analytics_interop.getAnalyticsInstance(core_interop.app(app.name));
+  }
+
+  /// Builds an instance of [FirebaseAnalyticsWeb] with an optional [FirebaseApp] instance
+  /// If [app] is null then the created instance will use the default [FirebaseApp]
+  FirebaseAnalyticsWeb({FirebaseApp? app}) : super(appInstance: app);
 
   /// Called by PluginRegistry to register this plugin for Flutter Web
   static void registerWith(Registrar registrar) {
@@ -19,36 +32,91 @@ class FirebaseAnalyticsWeb extends FirebaseAnalyticsPlatform {
   }
 
   @override
+  FirebaseAnalyticsPlatform delegateFor({FirebaseApp? app}) {
+    return FirebaseAnalyticsWeb(app: app);
+  }
+
+  @override
   Future<void> logEvent({
     required String name,
     Map<String, Object?>? parameters,
+    AnalyticsCallOptions? callOptions,
   }) async {
-    _analytics.logEvent(name, parameters ?? {});
+    return guard(() {
+      return _delegate.logEvent(
+        name: name,
+        parameters: parameters ?? {},
+        callOptions: callOptions,
+      );
+    });
+  }
+
+  @override
+  Future<void> setConsent({
+    bool? adStorageConsentGranted,
+    bool? analyticsStorageConsentGranted,
+  }) async {
+    throw UnimplementedError('setConsent() is not supported on Web.');
   }
 
   @override
   Future<void> setAnalyticsCollectionEnabled(bool enabled) async {
-    _analytics.setAnalyticsCollectionEnabled(enabled);
+    return guard(() {
+      return _delegate.setAnalyticsCollectionEnabled(enabled: enabled);
+    });
   }
 
   @override
-  Future<void> setUserId(String? id) async {
-    _analytics.setUserId(id);
+  Future<void> setUserId({
+    String? id,
+    AnalyticsCallOptions? callOptions,
+  }) async {
+    return guard(() {
+      return _delegate.setUserId(
+        id: id,
+        callOptions: callOptions,
+      );
+    });
   }
 
   @override
   Future<void> setCurrentScreen({
-    required String? screenName,
+    String? screenName,
     String? screenClassOverride,
+    AnalyticsCallOptions? callOptions,
   }) async {
-    _analytics.setCurrentScreen(screenName);
+    return guard(() {
+      return _delegate.setCurrentScreen(
+        screenName: screenName,
+        callOptions: callOptions,
+      );
+    });
+  }
+
+  @override
+  Future<void> resetAnalyticsData() async {
+    throw UnimplementedError('resetAnalyticsData() is not supported on Web.');
   }
 
   @override
   Future<void> setUserProperty({
     required String name,
     required String? value,
+    AnalyticsCallOptions? callOptions,
   }) async {
-    _analytics.setUserProperties({name: value});
+    return guard(() {
+      return _delegate.setUserProperty(
+        name: name,
+        value: value,
+        callOptions: callOptions,
+      );
+    });
+  }
+
+  @override
+  Future<void> setSessionTimeoutDuration(Duration timeout) async {
+    throw UnimplementedError(
+      'setSessionTimeoutDuration() is not supported on Web.',
+    );
   }
 }