+++ packages/firebase_dynamic_links/android/src/main/java/io/flutter/plugins/firebasedynamiclinks/FirebaseDynamicLinksPlugin.java
@@ -24,7 +24,6 @@ import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugin.common.PluginRegistry.NewIntentListener;
-import io.flutter.plugin.common.PluginRegistry.Registrar;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -51,24 +50,6 @@ public class FirebaseDynamicLinksPlugin
    * <p>Use this when adding the plugin to your FlutterEngine
    */
   public FirebaseDynamicLinksPlugin() {}
-
-  /**
-   * Registers a plugin with the v1 embedding api {@code io.flutter.plugin.common}.
-   *
-   * <p>Calling this will register the plugin with the passed registrar. However, plugins
-   * initialized this way won't react to changes in activity or context.
-   *
-   * @param registrar attaches this plugin's {@link
-   *     io.flutter.plugin.common.MethodChannel.MethodCallHandler} to the registrar's {@link
-   *     io.flutter.plugin.common.BinaryMessenger}.
-   */
-  public static void registerWith(Registrar registrar) {
-    final MethodChannel channel = createChannel(registrar.messenger());
-    final FirebaseDynamicLinksPlugin plugin =
-        new FirebaseDynamicLinksPlugin(registrar.activity(), channel);
-    registrar.addNewIntentListener(plugin);
-    channel.setMethodCallHandler(plugin);
-  }
 
   private static MethodChannel createChannel(final BinaryMessenger messenger) {
     return new MethodChannel(messenger, "plugins.flutter.io/firebase_dynamic_links");