+++ packages/firebase_remote_config/firebase_remote_config/example/test_driver/firebase_remote_config_e2e.dart
@@ -1,23 +1,30 @@
 // ignore_for_file: require_trailing_commas
 // @dart = 2.9
 import 'package:drive/drive.dart' as drive;
-import 'package:flutter_test/flutter_test.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_remote_config/firebase_remote_config.dart';
+import 'package:flutter_test/flutter_test.dart';
 
 void testsMain() {
   group('RemoteConfig', () {
     RemoteConfig remoteConfig;
 
     setUp(() async {
-      await Firebase.initializeApp();
+      await Firebase.initializeApp(
+          options: const FirebaseOptions(
+        apiKey: 'AIzaSyAHAsf51D0A407EklG1bs-5wA7EbyfNFg0',
+        appId: '1:448618578101:ios:2bc5c1fe2ec336f8ac3efc',
+        storageBucket: 'react-native-firebase-testing.appspot.com',
+        databaseURL: 'https://react-native-firebase-testing.firebaseio.com',
+        messagingSenderId: '448618578101',
+        projectId: 'react-native-firebase-testing',
+      ));
       remoteConfig = RemoteConfig.instance;
       await remoteConfig.setConfigSettings(RemoteConfigSettings(
         fetchTimeout: const Duration(seconds: 8),
         minimumFetchInterval: Duration.zero,
       ));
       await remoteConfig.setDefaults(<String, dynamic>{
-        'welcome': 'default welcome',
         'hello': 'default hello',
       });
       await remoteConfig.ensureInitialized();
@@ -29,8 +36,8 @@ void testsMain() {
       await remoteConfig.fetchAndActivate();
       expect(remoteConfig.lastFetchStatus, RemoteConfigFetchStatus.success);
       expect(remoteConfig.lastFetchTime.isAfter(mark), true);
-      expect(remoteConfig.getString('welcome'), 'Earth, welcome! Hello!');
-      expect(remoteConfig.getValue('welcome').source, ValueSource.valueRemote);
+      expect(remoteConfig.getString('string'), 'invertase');
+      expect(remoteConfig.getValue('string').source, ValueSource.valueRemote);
 
       expect(remoteConfig.getString('hello'), 'default hello');
       expect(remoteConfig.getValue('hello').source, ValueSource.valueDefault);
@@ -54,7 +61,7 @@ void testsMain() {
       expect(remoteConfig.settings.minimumFetchInterval,
           const Duration(seconds: 88));
       await remoteConfig.setConfigSettings(RemoteConfigSettings(
-          fetchTimeout: const Duration(seconds: 10, milliseconds: 500),
+          fetchTimeout: const Duration(seconds: 10),
           minimumFetchInterval: Duration.zero));
       expect(remoteConfig.settings.fetchTimeout, const Duration(seconds: 10));
       expect(remoteConfig.settings.minimumFetchInterval, Duration.zero);