+++ packages/cloud_functions/cloud_functions_platform_interface/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 5.0.14
+
+ - Update a dependency to the latest release.
+
+## 5.0.13
+
+ - Update a dependency to the latest release.
+
+## 5.0.12
+
+ - Update a dependency to the latest release.
+
 ## 5.0.11
 
  - Update a dependency to the latest release.