+++ packages/cloud_functions/cloud_functions_web/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 4.1.1
+
+ - **FIX**: correctly pass `region` to JS functions interop instance (#7328).
+
+## 4.1.0
+
+ - **FEAT**: support initializing default `FirebaseApp` instances from Dart (#6549).
+
+## 4.0.15
+
+ - Update a dependency to the latest release.
+
+## 4.0.14
+
+ - Update a dependency to the latest release.
+
 ## 4.0.13
 
  - Update a dependency to the latest release.