+++ packages/firebase_database/firebase_database_web/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.1.1
+
+ - **FEAT**: support initializing default `FirebaseApp` instances from Dart (#6549).
+
+## 0.1.0+2
+
+ - Update a dependency to the latest release.
+
+## 0.1.0+1
+
+ - Update a dependency to the latest release.
+
 ## 0.1.0
 
 > Note: This release has breaking changes.