+++ packages/firebase_ml_custom/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0+4
+
+ - **REFACTOR**: remove deprecated Flutter Android v1 Embedding usages, including in example app (#7158).
+ - **CHORE**: update Gradle versions used in Android example app (#7054).
+
 ## 0.2.0+3
 
  - **STYLE**: enable additional lint rules (#6832).