+++ packages/firebase_ml_vision/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.12.0+3
+
+ - **STYLE**: enable additional lint rules (#6832).
+ - **REFACTOR**: remove deprecated Flutter Android v1 Embedding usages, including in example app (#7158).
+
 ## 0.12.0+2
 
  - **DOCS**: add discontinuation message and list alternatives (#6100).