+++ packages/firebase_remote_config/firebase_remote_config_platform_interface/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 1.0.0-dev.2
+
+ - Update a dependency to the latest release.
+
+## 1.0.0-dev.1
+
+ - Update a dependency to the latest release.
+
+## 1.0.0-dev.0
+
+ - Bump "firebase_remote_config_platform_interface" to `1.0.0-dev.0`.
+
+## 0.3.0+7
+
+ - Update a dependency to the latest release.
+
 ## 0.3.0+6
 
  - Update a dependency to the latest release.